+++ packages/react/src/components/InteractiveList/src/components/InteractiveListCheckboxItem/interactiveListCheckboxItem.types.ts
@@ -13,5 +13,5 @@ export interface InteractiveListCheckboxItemProperties
 
 export type InteractiveListCheckboxItemProps =
   InteractiveListCheckboxItemProperties &
-    Omit<HTMLAttributes<HTMLElement>, "color">&
+    Omit<HTMLAttributes<HTMLElement>, "color"> &
     Pick<InteractiveListRowProps, "topDivider">;