+++ packages/react/src/components/Menu/src/menu.spec.tsx
@@ -6,7 +6,7 @@ import { MenuProps } from "./menu.types";
 
 const bodyChildren = "Body content";
 
-const makeSut = (rest: Omit<MenuProps, "children">) => {
+const makeSut = (rest?: Omit<MenuProps, "children">) => {
   render(
     <Menu {...rest} data-testid="menu-element">
       {bodyChildren}
@@ -17,7 +17,7 @@ const makeSut = (rest: Omit<MenuProps, "children">) => {
 describe("GIVEN <Menu />", () => {
   describe("WHEN rendered", () => {
     it("SHOULD render children correctly", () => {
-      makeSut({});
+      makeSut();
       expect(screen.getByText(bodyChildren)).toBeDefined();
     });
   });