+++ packages/react/src/components/Page/src/components/PageBody/pageBody.types.ts
@@ -11,4 +11,5 @@ export interface PageBodyProperties {
 }
 
 export type PageBodyProps = PageBodyProperties &
-  Omit<HTMLAttributes<HTMLElement>, "color"> & BoxProperties;
+  Omit<HTMLAttributes<HTMLElement>, "color"> &
+  BoxProperties;