+++ packages/react/src/components/MenuButton/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 MenuButton component is meant to be used inside Menus and navigation, and allows the user to navigate between different sections of an application.
 
+## 2023-04-12 `1.3.0`
+
+#### 🎉 New features
+
+- Added `MenuButton.Accordion` subcomponent. ([#44](https://github.com/TiendaNube/nimbus-patterns/pull/44) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-04-04 `1.2.0`
 
 #### 🎉 New features